+++ src/components/zoneOutput.h
@@ -28,7 +28,7 @@ class zoneOutput{
         void close();
 
     private:
-        int _pin;
+        uint8_t _pin;
         
         
 };
